requirements.txt
django>=1.4,
south,
feincms-groupagenda,